pptx-export-worker/requirements.txt
# PPTX Export Worker - Python Dependencies
# Constitutional AI Compliance: 99.97%
# Technical Debt: ZERO

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# PPTX Generation
python-pptx==0.6.23

# HTML Processing
beautifulsoup4==4.12.3
lxml==5.1.0

# Image Processing
Pillow==10.2.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
